fifth 1

+++ 0x15-file_io/1-create_file.c
@@ -16,7 +16,7 @@ int create_file(const char *filename, char *text_content)
 {
 	int descriptor;
 	ssize_t num_of_bytes = 0;
-	ssize_t length = strlen(text_content);
+	int length = strlen(text_content);
 
 	if (!filename)
 	{
